[#9] .Add example using Cosmos
